Adds test for time conversions.
It is even worse, than I had thought.
If you ever end up fixing these tests, please feel free
to adopt the time conversion checks. These
rounding errors are sooo bad.

+++ include/otf2xx/chrono/convert.hpp
@@ -134,7 +134,7 @@ namespace chrono
          */
         otf2::chrono::ticks operator()(time_point t) const
         {
-            auto tp = t.time_since_epoch().count();
+            auto tp = static_cast<uint64_t>(t.time_since_epoch().count());
 
             assert(tp <
                    static_cast<uint64_t>(std::numeric_limits<int64_t>::max()) / inverse_factor_);
